+++ src/ODE.jl
@@ -10,22 +10,28 @@ using ForwardDiff
 import Base.convert, Base.show
 import Base: start, next, done, call, collect
 
-# basic type definitions
+# Core infrastructure
+#
+# When wrapping a new solver it will need to use and conform to
+# methods and types within these files.
+#
+# Note, if we go the MathProgBase.jl route, then these files would go
+# into ODEBase.jl.
 include("types.jl")
+include("tableaus.jl")
 include("options.jl")
 include("helpers.jl")
 
-# dense output wrapper
+# Dense output wrapper
 include("dense.jl")
 
-# particular solvers
+# Particular solvers
 include("ode23s.jl")
 include("runge-kutta.jl")
-# include("multistep.jl")
 include("adams-bashford-moulton.jl")
 include("rosenbrock.jl")
-# include("taylor.jl")
 
+# User interface to solvers
 include("interfaces.jl")
 
 end # module ODE
